node target get explorer url function migration
